+++ hoomd/hpmc/module.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2009-2023 The Regents of the University of Michigan.
+// Copyright (c) 2009-2024 The Regents of the University of Michigan.
 // Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 // Include the defined classes that are to be exported to python
@@ -46,9 +46,9 @@ void exportPairPotential(pybind11::module& m);
 
 void exportPairPotentialLennardJones(pybind11::module& m);
 
-void export_PairPotentialUnion(pybind11::module& m);
-
 void exportPairPotentialAngularStep(pybind11::module& m);
+
+void exportPairPotentialUnion(pybind11::module& m);
     } // namespace detail
     } // namespace hpmc
     } // namespace hoomd
@@ -146,8 +146,8 @@ PYBIND11_MODULE(_hpmc, m)
 
     exportPairPotential(m);
     exportPairPotentialLennardJones(m);
-    export_PairPotentialUnion(m);
     exportPairPotentialAngularStep(m);
+    exportPairPotentialUnion(m);
     }
 
 /*! \defgroup hpmc_integrators HPMC integrators