+++ hoomd/hpmc/pytest/test_pair_union.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2009-2023 The Regents of the University of Michigan.
+# Copyright (c) 2009-2024 The Regents of the University of Michigan.
 # Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 """Test hoomd.hpmc.pair.Union."""
@@ -67,9 +67,9 @@ def pair_union_simulation_factory(simulation_factory,
                                   two_particle_snapshot_factory):
     """Make two particle sphere simulations with a union potential."""
 
-    def make_union_sim(union_potential, particle_types=['A'], d=1):
+    def make_union_sim(union_potential, particle_types=['A'], d=1, L=20):
         sim = simulation_factory(
-            two_particle_snapshot_factory(particle_types, d=d))
+            two_particle_snapshot_factory(particle_types, d=d, L=L))
         sphere = hpmc.integrate.Sphere()
         sphere.shape["A"] = dict(diameter=2.0)
         sphere.pair_potentials = [union_potential]
@@ -159,8 +159,8 @@ def test_get_set_body_params(pair_union_simulation_factory, union_potential):
     union_potential.body["A"] = body_dict
     assert union_potential.body["A"]["positions"] == body_dict["positions"]
     assert union_potential.body["A"]["types"] == body_dict["types"]
-    assert union_potential.body["A"]["orientations"] is None
-    assert union_potential.body["A"]["charges"] is None
+    assert 'orientations' not in union_potential.body["A"]
+    assert 'charges' not in union_potential.body["A"]
 
     # after attaching, setting as dict
     sim = pair_union_simulation_factory(union_potential)
@@ -187,7 +187,33 @@ def test_get_set_body_params(pair_union_simulation_factory, union_potential):
     assert union_potential.body["A"]["charges"] == [0]
 
 
-# TODO test get/set leaf capacity and constituent potential properties
+@pytest.mark.cpu
+def test_get_set_properties(pair_union_simulation_factory, union_potential):
+    """Test getting/setting leaf capacity and constituent potential."""
+    # assert values are right on construction
+    assert union_potential.leaf_capacity == 0
+    lj = union_potential.constituent_potential
+    assert lj.params[('A', 'A')] == dict(epsilon=1.0,
+                                         sigma=1.0,
+                                         r_cut=2.0,
+                                         r_on=0.0)
+
+    # try to set params
+    lj2 = hpmc.pair.LennardJones()
+    lj2.params[('A', 'A')] = dict(epsilon=0.5, sigma=2.0, r_cut=3.0)
+    with pytest.raises(AttributeError):
+        union_potential.constituent_potential = lj2
+    union_potential.leaf_capacity = 3
+    assert union_potential.leaf_capacity == 3
+
+    # attach
+    union_potential.body["A"] = dict(types=["A"], positions=[(0, 0, 1)])
+    sim = pair_union_simulation_factory(union_potential)
+    sim.run(0)
+
+    # set after attaching
+    union_potential.leaf_capacity = 5
+    assert union_potential.leaf_capacity == 5
 
 
 @pytest.mark.cpu
@@ -221,7 +247,8 @@ def test_energy(pair_union_simulation_factory, union_potential, leaf_capacity):
                                      positions=[(-1, 0, 0), (1, 0, 0)])
     sim = pair_union_simulation_factory(union_potential,
                                         particle_types=['A', 'B'],
-                                        d=3)
+                                        d=3,
+                                        L=30)
     sim.operations.integrator.shape["B"] = dict(diameter=2)
     sim.run(0)