+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-.. Copyright (c) 2009-2023 The Regents of the University of Michigan.
+.. Copyright (c) 2009-2024 The Regents of the University of Michigan.
 .. Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 Change Log
@@ -7,7 +7,7 @@ Change Log
 4.x
 ---
 
-4.5.0 (2023-02-13)
+4.5.0 (2024-02-13)
 ^^^^^^^^^^^^^^^^^^
 
 *Fixed*
@@ -61,6 +61,8 @@ Change Log
   (`#1676 <https://github.com/glotzerlab/hoomd-blue/pull/1676>`__).
 * ``HPMCIntegrator.pair_potential`` - Use compiled potentials with ``pair_potentials``
   (`#1676 <https://github.com/glotzerlab/hoomd-blue/pull/1676>`__).
+* Single-process multi-gpu code path
+  (`#1706 <https://github.com/glotzerlab/hoomd-blue/pull/1706>`__).
 
 *Changed*